+++ rich/progress.py
@@ -48,7 +48,6 @@ TaskID = NewType("TaskID", int)
 
 ProgressType = TypeVar("ProgressType")
 
-
 GetTimeCallable = Callable[[], float]
 
 
@@ -293,15 +292,30 @@ class TotalFileSizeColumn(ProgressColumn):
 
 
 class DownloadColumn(ProgressColumn):
-    """Renders file size downloaded and total, e.g. '0.5/2.3 GB'."""
+    """Renders file size downloaded and total, e.g. '0.5/2.3 GB'.
+
+    Args:
+        binary_units (bool, optional): Flag to render filesize in binary format, enable to render in binary. Defaults to False.
+    """
+
+    def __init__(self, binary_units: bool = False) -> None:
+        self.binary_units = binary_units
+        super().__init__()
 
     def render(self, task: "Task") -> Text:
         """Calculate common unit for completed and total."""
         completed = int(task.completed)
         total = int(task.total)
-        unit, suffix = filesize.pick_unit_and_suffix(
-            total, ["bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"], 1000
-        )
+        if self.binary_units:
+            unit, suffix = filesize.pick_unit_and_suffix(
+                total,
+                ["bytes", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB"],
+                1024,
+            )
+        else:
+            unit, suffix = filesize.pick_unit_and_suffix(
+                total, ["bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"], 1000
+            )
         completed_ratio = completed / unit
         total_ratio = total / unit
         precision = 0 if unit == 1 else 1