+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added highlighting of EUI-48 and EUI-64 (MAC addresses)
 - Added Console.pager
 - Added Console.out
-- Added decimal_suffix in progress download column
+- Added binary_units in progress download column
 
 ### Changed